tests/submission/Microsoft/forms/inputtypes.html

<!DOCTYPE html>
<html><head>
        <title>email type on the input element</title>
        <meta content="text/html; charset=UTF-8" http-equiv="content-type">
        <meta content="Input types on elements" name="description">
    <link href="http://www.w3.org/TR/html5/the-input-element.html#dom-input-type" rel="help">
</head>
    <body>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>

        <div id="log"></div> 
        
        <div style="display : none">
            <input type="email">

            <datalist id="datalist1">
              <option>1</option>
              <option>2</option>
              <option>3</option>
              <option>4</option>
            </datalist>
            <input list="datalist1" type="text">
            <input type="number">
        </div>
        
        <script type="text/javascript">

        test(function() {assert_equals(document.getElementsByTagName("input")[0].type, "email")}, "email type support on input element");
        test(function() {assert_equals(document.getElementsByTagName("input")[1].list, document.getElementsByTagName("datalist")[0])}, "list attribute on the input element", {
          "help": "http://www.w3.org/TR/html5/common-input-element-apis.html#dom-input-list"
        });
        test(function() {assert_equals(document.getElementsByTagName("input")[2].type, "number")}, "number type support on input element");
 
        
        
        </script>
    
</body></html>
